+++ Quiper/SettingsView.swift
@@ -179,9 +179,9 @@ struct GeneralSettingsView: View {
                     }
                 }
             }
+            .padding(16)
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .background(Color(NSColor.windowBackgroundColor))
         .background(Color(NSColor.windowBackgroundColor))
         .onAppear {
             launchAtLogin = Launcher.isInstalledAtLogin()